refactor: Mv getClass and create to helpers
