+++ packages/cozy-client/src/associations/helpers.js
@@ -1,6 +1,8 @@
-import Association from './Association'
 import pick from 'lodash/pick'
 import pickBy from 'lodash/pickBy'
+import Association from './Association'
+import HasMany from './HasManyAssociation'
+import HasManyFiles from './HasManyFilesAssociation'
 
 export const pickTypeAndId = x => pick(x, '_type', '_id')
 const applyHelper = (fn, objOrArr) =>
@@ -13,3 +15,31 @@ export const responseToRelationship = response =>
     next: response.next,
     skip: response.skip
   })
+
+export const getClass = (doctype, type) => {
+  if (typeof type !== 'string') {
+    return type
+  } else {
+    switch (type) {
+      case 'has-many':
+        return doctype === 'io.cozy.files'
+          ? HasManyFilesAssociation
+          : HasManyAssociation
+    }
+  }
+  throw new Error(`Unknown association '${type}'`)
+}
+
+export const create = (
+  target,
+  { name, type, doctype },
+  { get, query, mutate, save }
+) => {
+  const accessors = { get, query, mutate, save }
+
+  if (target[name] instanceof Association) {
+    throw new Error(`Association ${name} already exists`)
+  }
+
+  return new type(target, name, doctype, accessors)
+}
